FourthLab with sorting methods in FilmMakingMangerUtils class, using interfaces Comparable, Comparator and lambda expressions
