+++ filmEquipment/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "java.configuration.updateBuildConfiguration": "automatic",
+    "editor.wordWrapColumn": 150,
+    "html.format.wrapLineLength": 100,
+    "editor.wrappingIndent": "deepIndent"
+}
