+++ filmEquipment/src/main/java/ua/lviv/iot/equipment/model/AbstractRecordingDevice.java
@@ -1,0 +1,36 @@
+package ua.lviv.iot.equipment.model;
+
+public abstract class AbstractRecordingDevice extends AbstarctShootingEquipment {
+
+  private boolean isWaterproof;
+  private int batteryLifeInHours;
+  private String recordFormat;
+
+  public AbstractRecordingDevice(int warrantyWorkPeriodInMonths, double weightInGrams) {
+    super(warrantyWorkPeriodInMonths, weightInGrams);
+  }
+
+  public boolean isWaterproof() {
+    return isWaterproof;
+  }
+
+  public void setWaterproof(boolean isWaterproof) {
+    this.isWaterproof = isWaterproof;
+  }
+
+  public int getBatteryLifeInHours() {
+    return batteryLifeInHours;
+  }
+
+  public void setBatteryLifeInHours(int batteryLifeInHours) {
+    this.batteryLifeInHours = batteryLifeInHours;
+  }
+
+  public String getRecordFormat() {
+    return recordFormat;
+  }
+
+  public void setRecordFormat(String recordFormat) {
+    this.recordFormat = recordFormat;
+  }
+}
