+++ filmEquipment/src/main/java/ua/lviv/iot/equipment/model/Quality.java
@@ -1,0 +1,5 @@
+package ua.lviv.iot.equipment.model;
+
+public enum Quality {
+  LOW, MIDDLE, HIGH
+}
