--- filmEquipment/src/main/java/ua/lviv/iot/thirdLab/filmMaking/manager/FilmMakingManagerUtils.java
@@ -1,17 +1,0 @@
-package ua.lviv.iot.thirdLab.filmMaking.manager;
-
-import java.util.List;
-
-import ua.lviv.iot.thirdLab.filmMaking.model.Costume;
-import ua.lviv.iot.thirdLab.filmMaking.model.SortType;
-
-public class FilmMakingManagerUtils {
-
-	public static void sortCostumesByYearOfUse(List<Costume> costumes, SortType type) {
-		
-	}
-
-	public static void sortCostumesByProductionYear(List<Costume> costumes, SortType type) {
-		
-	}
-}
