--- filmEquipment/src/main/java/ua/lviv/iot/thirdLab/filmMaking/model/Quality.java
@@ -1,5 +1,0 @@
-package ua.lviv.iot.thirdLab.filmMaking.model;
-
-public enum Quality {
-	LOW, MIDDLE, HIGH;
-}
