--- filmEquipment/src/main/java/ua/lviv/iot/thirdLab/filmMaking/model/SortType.java
@@ -1,5 +1,0 @@
-package ua.lviv.iot.thirdLab.filmMaking.model;
-
-public enum SortType {
-	ASC, DESC;
-}
